Observations 1, 2, 3, 5, fixed
